+++ src/components/AddButton/AddButton.style.ts
@@ -1,0 +1,8 @@
+import { css } from '@emotion/css';
+
+export const addButton = css`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 4px;
+`;
