+++ src/components/DeleteButton/DeleteButton.style.ts
@@ -1,0 +1,8 @@
+import { css } from '@emotion/css';
+
+export const deleteButton = css`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 4px;
+`;
