+++ src/components/Product/Product.tsx
@@ -1,43 +1,14 @@
-import { css } from '@emotion/css';
 import AddButton from '../AddButton/AddButton';
 import DeleteButton from '../DeleteButton/DeleteButton';
 import { ProductProps } from '../../types/product';
-
-const productContainer = css`
-  width: 100%;
-  height: 224px;
-  border-radius: 8px;
-  position: relative;
-`;
-
-const productImage = css`
-  width: 100%;
-  height: 50%;
-  border-top-left-radius: 4px;
-  border-top-right-radius: 4px;
-`;
-
-const productContent = css`
-  padding: 10px;
-`;
-
-const productTitle = css`
-  font-size: 14px;
-  font-weight: 700;
-  line-height: 20px;
-`;
-
-const productPrice = css`
-  font-size: 12px;
-  font-weight: 500;
-  line-height: 15px;
-`;
-
-const buttonWrapper = css`
-  position: absolute;
-  bottom: 10px;
-  right: 10px;
-`;
+import {
+  productContainer,
+  productImage,
+  productContent,
+  productTitle,
+  productPrice,
+  buttonWrapper,
+} from './Product.style';
 
 function Product({
   id,
