+++ src/components/ProductList/ProductList.style.ts
@@ -1,0 +1,7 @@
+import { css } from '@emotion/css';
+
+export const ProductListContainer = css`
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 10px;
+`;
