+++ src/components/common/SelectBox/SelectBox.style.ts
@@ -1,0 +1,9 @@
+import { css } from '@emotion/css';
+
+export const selectBoxContainer = css`
+  width: 100%;
+  height: 36px;
+  padding: 0 10px;
+  border-radius: 8px;
+  border: 1px solid rgba(0, 0, 0, 0.1);
+`;
