+++ src/components/common/SelectBox/SelectBox.tsx
@@ -1,17 +1,9 @@
-import { css } from '@emotion/css';
+import { selectBoxContainer } from './SelectBox.style';
 
 type SelectBoxProps = {
   placeHolder: string;
   options: string[];
 } & React.SelectHTMLAttributes<HTMLElement>;
-
-const selectBoxContainer = css`
-  width: 100%;
-  height: 36px;
-  padding: 0 10px;
-  border-radius: 8px;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-`;
 
 function SelectBox({ placeHolder, options, ...props }: SelectBoxProps) {
   return (
